refactor(alerts): decouple editor from modal container
Move alert editor structure and behavior into container-agnostic primitives, keep LemonModal integration in a thin adapter, and cover both embedded and modal rendering in Storybook.

Generated-By: PostHog Code
Task-Id: e4383360-d5c1-4160-b581-3c5d787f236d
