+++ products/logs/frontend/components/LogsAlerting/LogsAlertForm.tsx
@@ -16,7 +16,7 @@ import { Tooltip } from 'lib/lemon-ui/Tooltip'
 
 import { AnyPropertyFilter, PropertyFilterType, PropertyOperator, UniversalFiltersGroup } from '~/types'
 
-import { AlertEditorFormDetails, AlertEditorSection } from 'products/alerts/frontend/components/AlertEditorModal'
+import { AlertEditorFormDetails, AlertEditorSection } from 'products/alerts/frontend/components/AlertEditor'
 import { ServiceFilter } from 'products/logs/frontend/components/LogsViewer/Filters/ServiceFilter'
 import { SeverityLevelsFilter } from 'products/logs/frontend/components/LogsViewer/Filters/SeverityLevelsFilter'
 import { LogsAlertThresholdOperatorEnumApi } from 'products/logs/frontend/generated/api.schemas'
